add snapshot time to packagemanager
