+++ toolkit/tools/pkg/imagecustomizerlib/customizepackages.go
@@ -24,12 +24,13 @@ var (
 	ErrPackageInstall             = NewImageCustomizerError("Packages:Install", "failed to install packages")
 	ErrPackageCacheClean          = NewImageCustomizerError("Packages:CacheClean", "failed to clean tdnf cache")
 	ErrMountRpmSources            = NewImageCustomizerError("Packages:MountRpmSources", "failed to mount RPM sources")
+	ErrSnapshotTimeNotSupported   = NewImageCustomizerError("Packages:SnapshotTimeNotSupported", "snapshot time is not supported")
 )
 
 // addRemoveAndUpdatePackages orchestrates the complete package management workflow
 func addRemoveAndUpdatePackages(ctx context.Context, buildDir string, baseConfigPath string, config *imagecustomizerapi.OS,
 	imageChroot *safechroot.Chroot, toolsChroot *safechroot.Chroot,
-	rpmsSources []string, useBaseImageRpmRepos bool, distroConfig distroHandler, snapshotTime string,
+	rpmsSources []string, useBaseImageRpmRepos bool, distroHandler distroHandler, snapshotTime string,
 ) error {
 	ctx, span := otel.GetTracerProvider().Tracer(OtelTracerName).Start(ctx, "configure_packages")
 	defer span.End()
@@ -39,7 +40,7 @@ func addRemoveAndUpdatePackages(ctx context.Context, buildDir string, baseConfig
 	}
 
 	// Delegate the entire package management workflow to the distribution-specific implementation
-	return distroConfig.managePackages(ctx, buildDir, baseConfigPath, config, imageChroot, toolsChroot, rpmsSources, useBaseImageRpmRepos, snapshotTime)
+	return distroHandler.managePackages(ctx, buildDir, baseConfigPath, config, imageChroot, toolsChroot, rpmsSources, useBaseImageRpmRepos, snapshotTime)
 }
 
 func collectPackagesList(baseConfigPath string, packageLists []string, packages []string) ([]string, error) {
