+++ toolkit/tools/pkg/imagecustomizerlib/packagemanager_rpm.go
@@ -3,20 +3,6 @@
 
 package imagecustomizerlib
 
-import (
-	"context"
-	"fmt"
-	"os"
-	"path/filepath"
-
-	"github.com/microsoft/azurelinux/toolkit/tools/internal/logger"
-	"github.com/microsoft/azurelinux/toolkit/tools/internal/safechroot"
-	"github.com/microsoft/azurelinux/toolkit/tools/internal/shell"
-	"github.com/sirupsen/logrus"
-	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/attribute"
-)
-
 // rpmPackageManagerHandler represents the interface for RPM-based package managers (TDNF, DNF)
 type rpmPackageManagerHandler interface {
 	// Package manager configuration
@@ -29,151 +15,7 @@ type rpmPackageManagerHandler interface {
 
 	// Package manager specific cache options for install/update operations
 	getCacheOnlyOptions() []string
-}
 
-// executeRpmPackageManagerCommand runs a package manager command with proper chroot handling
-func executeRpmPackageManagerCommand(args []string, imageChroot *safechroot.Chroot, toolsChroot *safechroot.Chroot, pmHandler rpmPackageManagerHandler) error {
-	pmChroot := imageChroot
-	if toolsChroot != nil {
-		pmChroot = toolsChroot
-	}
-
-	if _, err := os.Stat(filepath.Join(pmChroot.RootDir(), pmHandler.getConfigFile())); err == nil {
-		args = append([]string{"--config", "/" + pmHandler.getConfigFile()}, args...)
-	}
-
-	// Use package manager specific output callback
-	stdoutCallback := pmHandler.createOutputCallback()
-
-	return pmChroot.UnsafeRun(func() error {
-		return shell.NewExecBuilder(pmHandler.getPackageManagerBinary(), args...).
-			StdoutCallback(stdoutCallback).
-			LogLevel(shell.LogDisabledLevel, logrus.DebugLevel).
-			ErrorStderrLines(1).
-			Execute()
-	})
-}
-
-func installOrUpdateRpmPackages(ctx context.Context, action string, allPackagesToAdd []string, imageChroot *safechroot.Chroot, toolsChroot *safechroot.Chroot, pmHandler rpmPackageManagerHandler) error {
-	if len(allPackagesToAdd) == 0 {
-		return nil
-	}
-
-	_, span := otel.GetTracerProvider().Tracer(OtelTracerName).Start(ctx, action+"_packages")
-	span.SetAttributes(
-		attribute.Int(action+"_packages_count", len(allPackagesToAdd)),
-		attribute.StringSlice("packages", allPackagesToAdd),
-	)
-	defer span.End()
-
-	// Build command arguments directly
-	args := []string{"-v", action, "--assumeyes", "--cacheonly"}
-
-	args = append(args, "--setopt=reposdir="+rpmsMountParentDirInChroot) // Add package manager specific cache options (e.g., DNF cache metadata options)
-	cacheOptions := pmHandler.getCacheOnlyOptions()
-	args = append(args, cacheOptions...)
-
-	args = append(args, allPackagesToAdd...)
-
-	if toolsChroot != nil {
-		args = append([]string{"--releasever=" + pmHandler.getReleaseVersion(), "--installroot=/" + toolsRootImageDir}, args...)
-	}
-
-	err := executeRpmPackageManagerCommand(args, imageChroot, toolsChroot, pmHandler)
-	if err != nil {
-		if action == "install" {
-			return fmt.Errorf("%w (%v):\n%w", ErrPackageInstall, allPackagesToAdd, err)
-		} else {
-			return fmt.Errorf("%w (%v):\n%w", ErrPackageUpdate, allPackagesToAdd, err)
-		}
-	}
-	return nil
-}
-
-// updateAllRpmPackages updates all packages using the appropriate package manager
-func updateAllRpmPackages(ctx context.Context, imageChroot *safechroot.Chroot, toolsChroot *safechroot.Chroot, pmHandler rpmPackageManagerHandler) error {
-	_, span := otel.GetTracerProvider().Tracer(OtelTracerName).Start(ctx, "update_base_packages")
-	defer span.End()
-
-	// Build command arguments directly
-	args := []string{"-v", "update", "--assumeyes", "--cacheonly"}
-
-	args = append(args, "--setopt=reposdir="+rpmsMountParentDirInChroot) // Add package manager specific cache options (e.g., DNF cache metadata options)
-	cacheOptions := pmHandler.getCacheOnlyOptions()
-	args = append(args, cacheOptions...)
-
-	if toolsChroot != nil {
-		args = append([]string{"--releasever=" + pmHandler.getReleaseVersion(), "--installroot=/" + toolsRootImageDir}, args...)
-	}
-
-	err := executeRpmPackageManagerCommand(args, imageChroot, toolsChroot, pmHandler)
-	if err != nil {
-		return fmt.Errorf("%w:\n%w", ErrPackagesUpdateInstalled, err)
-	}
-	return nil
-}
-
-// removeRpmPackages removes packages using the appropriate package manager
-func removeRpmPackages(ctx context.Context, allPackagesToRemove []string, imageChroot *safechroot.Chroot, toolsChroot *safechroot.Chroot, pmHandler rpmPackageManagerHandler) error {
-	if len(allPackagesToRemove) <= 0 {
-		return nil
-	}
-
-	_, span := otel.GetTracerProvider().Tracer(OtelTracerName).Start(ctx, "remove_packages")
-	span.SetAttributes(
-		attribute.Int("remove_packages_count", len(allPackagesToRemove)),
-		attribute.StringSlice("remove_packages", allPackagesToRemove),
-	)
-	defer span.End()
-
-	// Build command arguments directly
-	args := []string{"-v", "remove", "--assumeyes", "--disablerepo", "*"}
-	args = append(args, allPackagesToRemove...)
-
-	if toolsChroot != nil {
-		args = append([]string{"--releasever=" + pmHandler.getReleaseVersion(), "--installroot=/" + toolsRootImageDir}, args...)
-	}
-
-	err := executeRpmPackageManagerCommand(args, imageChroot, toolsChroot, pmHandler)
-	if err != nil {
-		return fmt.Errorf("%w (%v):\n%w", ErrPackageRemove, allPackagesToRemove, err)
-	}
-	return nil
-}
-
-// refreshRpmPackageMetadata refreshes package metadata
-func refreshRpmPackageMetadata(ctx context.Context, imageChroot *safechroot.Chroot, toolsChroot *safechroot.Chroot, pmHandler rpmPackageManagerHandler) error {
-	_, span := otel.GetTracerProvider().Tracer(OtelTracerName).Start(ctx, "refresh_metadata")
-	defer span.End()
-
-	logger.Log.Infof("Refreshing package metadata")
-
-	args := []string{"-v", "check-update", "--refresh", "--assumeyes"}
-
-	args = append(args, "--setopt=reposdir="+rpmsMountParentDirInChroot)
-
-	if toolsChroot != nil {
-		args = append([]string{"--releasever=" + pmHandler.getReleaseVersion(), "--installroot=/" + toolsRootImageDir}, args...)
-	}
-
-	err := executeRpmPackageManagerCommand(args, imageChroot, toolsChroot, pmHandler)
-	if err != nil {
-		return fmt.Errorf("%w:\n%w", ErrPackageRepoMetadataRefresh, err)
-	}
-	return nil
-}
-
-func cleanRpmCache(imageChroot *safechroot.Chroot, toolsChroot *safechroot.Chroot, pmHandler rpmPackageManagerHandler) error {
-	// Build command arguments directly
-	args := []string{"-v", "clean", "all"}
-
-	if toolsChroot != nil {
-		args = append([]string{"--releasever=" + pmHandler.getReleaseVersion(), "--installroot=/" + toolsRootImageDir}, args...)
-	}
-
-	err := executeRpmPackageManagerCommand(args, imageChroot, toolsChroot, pmHandler)
-	if err != nil {
-		return fmt.Errorf("%w:\n%w", ErrPackageCacheClean, err)
-	}
-	return nil
+	// Package manager specific snapshot time support
+	supportsSnapshotTime() bool
 }
